Work with SQLite DB and binary atribut

+++ app.py
@@ -1,8 +1,8 @@
 from flask import Flask, render_template, request, send_file
 from flask_sqlalchemy import SQLAlchemy
 from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField
-from wtforms.validators import DataRequired
+from wtforms_alchemy.fields import QuerySelectField
+
 import os
 
 __author__ = 'Vladan M. Kostic'
@@ -19,14 +19,22 @@ app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 db = SQLAlchemy(app)
 
 class FileContents(db.Model):
-
     id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(128))
+    name_of_person = db.Column(db.String(64))
+    name_of_file = db.Column(db.String(128))
     data = db.Column(db.LargeBinary)
 
-class SlikaPregledForm(FlaskForm):
-    choice_of_image = StringField('Enter key of image:', validators=[DataRequired()])
-    submit = SubmitField('Show')
+    def __repr__(self):
+        return '{}'.format(self.name_of_file)
+
+def choice_query():
+    return FileContents.query.all()
+
+class ChiceImageForm(FlaskForm):
+    opts = QuerySelectField(query_factory=choice_query,  get_label='name_of_file', allow_blank=False)
+
+class ChicePersonForChangeImageForm(FlaskForm):
+    opts_id = QuerySelectField(query_factory=choice_query, get_label='id', allow_blank=False)
 
 @app.route('/')
 @app.route('/index')
@@ -46,66 +54,62 @@ def upload():
     if not os.path.isdir(target):
         os.mkdir(target)
 
+    person = request.form['person']
     file = request.files['file']
+    print(file, ' ', person)
 
-    for file in request.files.getlist("file"):
-        print(file)
-        filename = file.filename
-        destination = "/".join([target, filename])
-        print(destination)
-        file.save(destination)
+    filename = file.filename
+    destination = "/".join([target, filename])
+    print(destination)
+    file.save(destination)
 
-        newFile = FileContents(name=file.filename, data=file.read())
-        db.session.add(newFile)
-        db.session.commit()
+    thedata = open(destination, 'rb').read()
+    print(thedata)
+
+    newFile = FileContents(name_of_person=person, name_of_file=file.filename, data=thedata)
+    db.session.add(newFile)
+    db.session.commit()
 
     return render_template("complete.html", filename=file.filename)
 
 @app.route('/show', methods=['GET','POST'])
 def show():
-    search = SlikaPregledForm(request.form)
-    if request.method == 'POST':
-        return search_results_slika(search)
+    form = ChiceImageForm()
+    if form.validate_on_submit():
+        print(form.opts.data)
+        file = FileContents.query.filter_by(name_of_file=str(form.opts.data)).first()
+        return render_template('/download.html', file=file)
 
-    return render_template('show.html',title='Show', form=search)
-
-#app.route('/download')
-def search_results_slika(search):
-    file = FileContents.query.filter_by(id=search.data['choice_of_image']).first()
-    return render_template('/download.html', file=file)
-
-@app.route('/ch', methods=['GET','POST'])
-def ch():
-    return render_template('change.html', title='Change')
+    return render_template('show.html',title='Show', form=form)
 
 @app.route('/change', methods=['GET','POST'])
 def change():
-    target = os.path.join(app_root, 'static/images/')
-    print(target)
+    form = ChicePersonForChangeImageForm()
+    if form.validate_on_submit():
+        target = os.path.join(app_root, 'static/images/')
+        print(target)
 
-    if not os.path.isdir(target):
-        os.mkdir(target)
-
-    file = request.files['file']
-    old_id = request.form['img_id']
-    for file in request.files.getlist("file"):
+        file = request.files['file']
+        old_name_of_file = form.opts_id.data
         print(file)
         filename = file.filename
         destination = "/".join([target, filename])
         print(destination)
         file.save(destination)
+        print(old_name_of_file)
 
-    existFile = FileContents.query.filter_by(id = old_id).first()
+        existFile = FileContents.query.filter_by(name_of_file = str(old_name_of_file)).first()
+        print(existFile)
+        destination_del = "/".join([target, existFile.name_of_file])
+        print(destination_del)
+        os.remove(destination_del)
 
-    destination_del = "/".join([target, existFile.name])
-    print(destination_del)
-    os.remove(destination_del)
+        existFile.name_of_file = file.filename
+        existFile.data = file.read()
+        db.session.commit()
+        return render_template("complete_change.html", filename=file.filename)
 
-    existFile.name = file.filename
-    existFile.data = file.read()
-    db.session.commit()
-
-    return render_template("complete_change.html", filename=file.filename)
+    return render_template('change.html', title='Change', form=form)
 
 if __name__ =='__main__':
     app.run()
